requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
jinja2==3.1.2
python-multipart==0.0.6
python-dateutil==2.8.2

# adicionados:
python-dotenv==1.0.1
itsdangerous==2.2.0
passlib[bcrypt]==1.7.4
weasyprint>=61,<63
